latest game added to site

+++ jn-upload/jn-game-upload.sh
@@ -1,7 +1,20 @@
 #!/bin/bash
 
 cd ~
-echo "Jet-Net Build Your Own Game - webgame uploader."
+echo " "
+echo "==============================="
+echo "Jet-Net iPhone game link script"
+echo "==============================="
+echo " "
+
+# ask before proceeding
+sg=$(cat /usr/share/jetnet/dev/sg.id | tr -cd 123456789 | head -c 4)
+read -p "Enter the access code: " sgin
+if [ "$sg" != "$sgin" ]
+then
+    echo "INCORRECT"
+    exit;
+fi
 
 # get/generate the student's id for game
 date=$(date +%Y%m%d)
@@ -9,41 +22,54 @@ if [ -f studentid ]
 then
   studentid=$(cat studentid)
 else
-  studentid=$(cat /dev/urandom | tr -cd [:alnum:] | head -c 4)
+  studentid=$(cat /dev/urandom | tr -cd 23456789abcdefghjkmnpqrstuvwxyz | head -c 5)
   echo "$studentid" > studentid
+  chmod u-w studentid
 fi
+
 # set other variables
-ID="$date$studentid"
-GAMEURL="http://play.indiegamegarden.com/$ID"
-fn="$ID.zip"
-echo "File ID to be uploaded: $fn"
+GAMEURL="http://play.indiegamegarden.com/play.php?id=$date$studentid"
+UPLOADURL="http://play.indiegamegarden.com/j3t-upload.php?id=$date$studentid"
+TESTURL="http://play.indiegamegarden.com/j3t-active.php"
+fn="game.zip"
+
+# test server status
+curl $TESTURL -f -A "Mozilla/5.0 (Windows NT 6.1)" >& /dev/null
+if [ $? -ne 0 ]
+then
+    echo "Server not responding, try again later!"
+    echo " "
+    read -p "Press [ENTER] to close this window"
+    exit
+fi
 
 # put game html5 into a zip file
+echo "File for upload: $fn"
 rm -f $fn
 pushd ~/Desktop/Game/BreakIt/Export/html5/bin > /dev/null
 zip -q -r ~/$fn *
 popd > /dev/null
 
 # upload zip to server
-echo "Uploading file to web server... just a moment"
-curl 'http://play.indiegamegarden.com/jetnet-upload.php' \
-  -A "Mozilla/5.0 (Windows NT 6.1)" \
+echo "Uploading file ... just a moment"
+curl $UPLOADURL \
+  -f -A "Mozilla/5.0 (Windows NT 6.1)" \
   -F "fileToUpload=@$fn"
 
 if [ $? -eq 0 ]
 then
 
     # gen qr code
-    qrencode -s 6 -o ~/Desktop/QR.png $GAMEURL
-    eog -g ~/Desktop/QR.png & >& /dev/null
+    qrencode -s 6 -o ~/QR.png $GAMEURL
+    eog -g ~/QR.png & >& /dev/null
 
     echo " "
-    echo "------------------------------------------------"
-    echo "Your game is now online on:"
+    echo "----------------------------------------------------------------------"
+    echo "Your mobile game link is:"
     echo " $GAMEURL"
-    echo "------------------------------------------------"
+    echo "----------------------------------------------------------------------"
     echo " "
-    #zenity --info --text="Your game can be played at:\n\n$GAMEURL"
+    #zenity --info --text="blah:\n\n$GAMEURL"
 
 fi
 read -p "Press [ENTER] to close window"
